third_party/blink/web_tests/platform/mac-mac10.12/external/wpt/storage-access-api/storageAccess.testdriver.sub-expected.txt

This is a testharness.js-based test.
PASS Set up storage access rules
PASS [same-site] Cookie access is allowed: true
PASS [third-party-allowed-on-first-party-site] Cookie access is allowed: true
PASS [third-party-blocked-on-first-party-site] Cookie access is allowed: false
PASS [third-party-blocked-all] Cookie access is allowed: false
Harness: the test ran to completion.
